requirements.txt
kafka-python
pandas
pyarrow
fastparquet